Chore: Moved Container Exceptions for Exceptions folder
